fixed rendering bug and code cleanup
