+++ src/ce_render_fundementals.h
@@ -3,3 +3,4 @@
 #include "rendering/mesh.h"
 #include "rendering/shader.h"
 #include "rendering/texture.h"
+#include "rendering/vertex.h"
