+++ src/rendering/mesh.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "shader.h"
+#include "vertex.h"
 #include <GL/glew.h>
 #include <managers/asset_manager.h>
 
@@ -19,13 +20,13 @@ namespace ce {
 
 		void sendToShader(ce::Shader* shader, bool bind = true);
 
-		unsigned GetIndexCount() { return m_indexCount; };
+		size_t GetIndexCount() { return m_indexArraySize / sizeof(GLuint); };
 		void bind(bool VBO = true, bool EBO = true), unbind(bool VBO = true, bool EBO = true);
 
 	 private:
-		GLsizeiptr m_vertCount, m_indexCount;
+		GLsizeiptr m_vertArraySize, m_indexArraySize;
 		GLuint m_VAO, m_VBO, m_EBO;
 
-		void initVAO(Vertex* verts, unsigned* indices);
+		void initVAO(Vertex* verts, GLuint* indices);
 	};
 }
